Move vga definitions into header file
